feat: adding a feature flag for showing or hiding the sort bar - Adding tests for sort dropdown - refactoring sort bar into SortDropdown - refactoring styles into Sort.styles.tsx
