+++ src/components/results/Results.tsx
@@ -1,29 +1,29 @@
 import React, {useContext, useEffect, useState} from "react";
 import {Result, SortOptionType} from "../../types";
 import ResultCard from "./ResultCard";
-import sortListBy from "./sortListBy";
 import {FilterBar} from "./FilterBar";
-import Typography from "@mui/material/Typography";
 import {useLocation} from "react-router-dom";
 import Searching from "../../assets/Searching.png";
 import {GlobalStateContext} from "../../context/globalStates";
 import {grabCurrentFiltersFromURLParams} from "../../util/historyHelper";
 import {
   ListItem,
-  MatchSortContainer, ResultsList,
+  MatchSortContainer,
+  ResultsList,
   ResultsMatched,
   ResultsPage,
   RightSide,
-  SortContainer, StyledFormControl, StyledSelect,
-  StyledUL
+  StyledUL,
 } from "./Results.styles";
+import sortListBy from "./sort/sortListBy";
+import SortDropdown from "./sort/SortDropdown";
 
 const Results: React.FC = () => {
   const {
     filteredResults,
     setCurrentFilters,
     setFilteredResults,
-    setIsFilterOpen
+    setIsFilterOpen,
   } = useContext(GlobalStateContext);
 
   const [loading, setLoading] = useState(true);
@@ -45,17 +45,17 @@ const Results: React.FC = () => {
   }, [filteredResults]);
 
   useEffect(() => {
-    setFilteredResults((filteredResults) => sortListBy(filteredResults, sortOption));
+    setFilteredResults((filteredResults) =>
+      sortListBy(filteredResults, sortOption)
+    );
   }, [sortOption, setFilteredResults]);
 
   const renderResults = () => {
     if (filteredResults.length === 0) {
       return (
         <>
-          <img src={Searching} alt="No Results"/>
-          <p>
-            Try clearing some filters!
-          </p>
+          <img src={Searching} alt="No Results" />
+          <p>Try clearing some filters!</p>
         </>
       );
     }
@@ -78,39 +78,10 @@ const Results: React.FC = () => {
           <RightSide onClick={() => setIsFilterOpen(false)}>
             <MatchSortContainer>
               <ResultsMatched>{`${filteredResults.length} matches:`}</ResultsMatched>
-              <SortContainer>
-                <Typography>Sort by:</Typography>
-                <StyledFormControl>
-                  <StyledSelect
-                    defaultValue={SortOptionType.DueDateNewToOld}
-                    onChange={(event) =>
-                      setSortOption(event.target.value as SortOptionType)
-                    }
-                    native
-                    variant="standard"
-                    inputProps={{"aria-label": "sort by"}}
-                  >
-                    <option value={SortOptionType.InterestHighToLow}>
-                      Interest: High-Low
-                    </option>
-                    <option value={SortOptionType.InterestLowToHigh}>
-                      Interest: Low-High
-                    </option>
-                    <option value={SortOptionType.DueDateNewToOld}>
-                      Due date: New-Old
-                    </option>
-                    <option value={SortOptionType.DueDateOldToNew}>
-                      Due date: Old-New
-                    </option>
-                    <option value={SortOptionType.AwardAmountHighToLow}>
-                      Amount: High-Low
-                    </option>
-                    <option value={SortOptionType.AwardAmountLowToHigh}>
-                      Amount: Low-High
-                    </option>
-                  </StyledSelect>
-                </StyledFormControl>
-              </SortContainer>
+              {process.env.REACT_APP_SHOW_SORT &&
+                process.env.REACT_APP_SHOW_SORT === "true" && (
+                  <SortDropdown setSortOption={setSortOption} />
+                )}
             </MatchSortContainer>
             <ResultsList>{renderResults()}</ResultsList>
           </RightSide>
